fix: templates creation
