--- src/components/Layout/App/App.module.scss
@@ -1,40 +1,0 @@
-@import "@styles/global";
-
-.page {
-  opacity: 1;
-  width: 100%;
-  min-width: 1024px;
-}
-
-.authPage {
-  opacity: 1;
-  height: 100vh;
-  display: grid;
-  place-items: center;
-}
-
-.template {
-  display: grid;
-  grid-template-columns: 208px 1fr;
-  grid-template-areas:
-    "header header"
-    ". controls"
-    "navbar content";
-  grid-template-rows: auto;
-  gap: 0;
-}
-
-.loginTemplate{
-  background-color: $globalWhite;
-}
-
-.controls{
-  height: 76px;
-  border: 3px solid $lightGreenTag;
-  grid-area: controls;
-}
-
-.content{
-  border: 3px solid $lightGreenTag;
-  grid-area: content;
-}
