+++ src/components/Layout/App/App.tsx
@@ -1,41 +1,20 @@
-import { BrowserRouter, Route, Routes, useLocation } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Home from '@pages/Home';
 import NotFound from '@pages/NotFound';
 import Login from '@pages/Login/Login';
-import React, { useEffect, FC } from 'react';
-import Header from '@components/Layout/Header/Header';
-import Menu from '@components/Layout/Menu/Menu';
-import Loader from '@/components/ui/Loader/Loader';
-import { getProfileUser, logoutUser } from '@/services/redux/slices/auth/auth';
-import { useAppDispatch, useAppSelector } from '@/services/typeHooks';
-import { RootState } from '@/services/redux/store';
+import { useEffect } from 'react';
 import Logout from '@/pages/Logout';
-
-import styles from './App.module.scss';
+import Template from '@templates/Template/Template';
+import { useAppDispatch, useAppSelector } from '@services/typeHooks';
+import { RootState } from '@services/redux/store';
+import { getProfileUser, logoutUser } from '@services/redux/slices/auth/auth';
 
 const App = () => {
-  return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/login" element={<Login />} />
-      <Route path="/logout" element={<Logout />} />
-      <Route path="*" element={<NotFound />} />
-    </Routes>
-  );
-};
-
-interface TemplateProps {
-  children: React.ReactNode;
-}
-
-const Template: FC<TemplateProps> = ({ children }) => {
-  const location = useLocation();
   const dispatch = useAppDispatch();
-  const isLoginRoute = location.pathname === '/login';
+  const isLoading = useAppSelector((state: RootState) => state.user.isLoading);
   const isLoggedIn = useAppSelector(
     (state: RootState) => state.user.isLoggedIn
   );
-  const isLoading = useAppSelector((state: RootState) => state.user.isLoading);
   const access = localStorage.getItem('accessToken') ?? '';
 
   useEffect(() => {
@@ -46,29 +25,22 @@ const Template: FC<TemplateProps> = ({ children }) => {
     }
   }, [access, dispatch]);
 
-  if (isLoading) {
-    return (
-      <section className={isLoggedIn ? styles.page : styles.authPage}>
-        <Loader />
-      </section>
-    );
-  }
-
   return (
-    <div className={isLoginRoute ? styles.loginTemplate : styles.template}>
-      {!isLoginRoute && <Header />}
-      {!isLoginRoute && <Menu />}
-      {children}
-    </div>
+    <Template isLoading={isLoading} isLoggedIn={isLoggedIn}>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/logout" element={<Logout />} />
+        <Route path="*" element={<NotFound />} />
+      </Routes>
+    </Template>
   );
 };
 
 const WrappedApp = () => {
   return (
     <BrowserRouter>
-      <Template>
-        <App />
-      </Template>
+      <App />
     </BrowserRouter>
   );
 };
