+++ src/templates/LoginTemplate/LoginTemplate.module.scss
@@ -1,0 +1,5 @@
+@import "@styles/global";
+
+.template{
+  background-color: $globalWhite;
+}
